Collect memory-related tests on SQLListener in the same suite
